+++ spec/api-net-spec.js
@@ -61,703 +61,6 @@ describe('net module', () => {
     })
   })
 
-  describe('ClientRequest API', () => {
-    afterEach(() => {
-      session.defaultSession.webRequest.onBeforeRequest(null)
-    })
-
-    it('it should be able to abort an HTTP request after request end and before response', (done) => {
-      const requestUrl = '/requestUrl'
-      let requestReceivedByServer = false
-      server.on('request', (request, response) => {
-        switch (request.url) {
-          case requestUrl:
-            requestReceivedByServer = true
-            cancelRequest()
-            process.nextTick(() => {
-              response.statusCode = 200
-              response.statusMessage = 'OK'
-              response.end()
-            })
-            break
-          default:
-            handleUnexpectedURL(request, response)
-        }
-      })
-
-      let requestAbortEventEmitted = false
-      let requestFinishEventEmitted = false
-      let requestCloseEventEmitted = false
-
-      const urlRequest = net.request({
-        method: 'GET',
-        url: `${server.url}${requestUrl}`
-      })
-      urlRequest.on('response', (response) => {
-        expect.fail('Unexpected response event')
-      })
-      urlRequest.on('finish', () => {
-        requestFinishEventEmitted = true
-      })
-      urlRequest.on('error', () => {
-        expect.fail('Unexpected error event')
-      })
-      urlRequest.on('abort', () => {
-        requestAbortEventEmitted = true
-      })
-      urlRequest.on('close', () => {
-        requestCloseEventEmitted = true
-        expect(requestFinishEventEmitted).to.be.true()
-        expect(requestReceivedByServer).to.be.true()
-        expect(requestAbortEventEmitted).to.be.true()
-        expect(requestCloseEventEmitted).to.be.true()
-        done()
-      })
-
-      urlRequest.end(randomString(kOneKiloByte))
-      function cancelRequest () {
-        urlRequest.abort()
-      }
-    })
-
-    it('it should be able to abort an HTTP request after response start', (done) => {
-      const requestUrl = '/requestUrl'
-      let requestReceivedByServer = false
-      server.on('request', (request, response) => {
-        switch (request.url) {
-          case requestUrl:
-            requestReceivedByServer = true
-            response.statusCode = 200
-            response.statusMessage = 'OK'
-            response.write(randomString(kOneKiloByte))
-            break
-          default:
-            handleUnexpectedURL(request, response)
-        }
-      })
-
-      let requestFinishEventEmitted = false
-      let requestResponseEventEmitted = false
-      let requestAbortEventEmitted = false
-      let requestCloseEventEmitted = false
-      let responseAbortedEventEmitted = false
-
-      const urlRequest = net.request({
-        method: 'GET',
-        url: `${server.url}${requestUrl}`
-      })
-      urlRequest.on('response', (response) => {
-        requestResponseEventEmitted = true
-        const statusCode = response.statusCode
-        expect(statusCode).to.equal(200)
-        response.pause()
-        response.on('data', (chunk) => {
-        })
-        response.on('end', () => {
-          expect.fail('Unexpected end event')
-        })
-        response.resume()
-        response.on('error', () => {
-          expect.fail('Unexpected error event')
-        })
-        response.on('aborted', () => {
-          responseAbortedEventEmitted = true
-        })
-        urlRequest.abort()
-      })
-      urlRequest.on('finish', () => {
-        requestFinishEventEmitted = true
-      })
-      urlRequest.on('error', () => {
-        expect.fail('Unexpected error event')
-      })
-      urlRequest.on('abort', () => {
-        requestAbortEventEmitted = true
-      })
-      urlRequest.on('close', () => {
-        requestCloseEventEmitted = true
-        expect(requestFinishEventEmitted).to.be.true('request should emit "finish" event')
-        expect(requestReceivedByServer).to.be.true('request should be received by the server')
-        expect(requestResponseEventEmitted).to.be.true('"response" event should be emitted')
-        expect(requestAbortEventEmitted).to.be.true('request should emit "abort" event')
-        expect(responseAbortedEventEmitted).to.be.true('response should emit "aborted" event')
-        expect(requestCloseEventEmitted).to.be.true('request should emit "close" event')
-        done()
-      })
-      urlRequest.end(randomString(kOneKiloByte))
-    })
-
-    it('abort event should be emitted at most once', (done) => {
-      const requestUrl = '/requestUrl'
-      let requestReceivedByServer = false
-      server.on('request', (request, response) => {
-        switch (request.url) {
-          case requestUrl:
-            requestReceivedByServer = true
-            cancelRequest()
-            break
-          default:
-            handleUnexpectedURL(request, response)
-        }
-      })
-
-      let requestFinishEventEmitted = false
-      let requestAbortEventCount = 0
-      let requestCloseEventEmitted = false
-
-      const urlRequest = net.request({
-        method: 'GET',
-        url: `${server.url}${requestUrl}`
-      })
-      urlRequest.on('response', () => {
-        expect.fail('Unexpected response event')
-      })
-      urlRequest.on('finish', () => {
-        requestFinishEventEmitted = true
-      })
-      urlRequest.on('error', () => {
-        expect.fail('Unexpected error event')
-      })
-      urlRequest.on('abort', () => {
-        ++requestAbortEventCount
-        urlRequest.abort()
-      })
-      urlRequest.on('close', () => {
-        requestCloseEventEmitted = true
-        // Let all pending async events to be emitted
-        setTimeout(() => {
-          expect(requestFinishEventEmitted).to.be.true()
-          expect(requestReceivedByServer).to.be.true()
-          expect(requestAbortEventCount).to.equal(1)
-          expect(requestCloseEventEmitted).to.be.true()
-          done()
-        }, 500)
-      })
-
-      urlRequest.end(randomString(kOneKiloByte))
-      function cancelRequest () {
-        urlRequest.abort()
-        urlRequest.abort()
-      }
-    })
-
-    it('Requests should be intercepted by webRequest module', (done) => {
-      const requestUrl = '/requestUrl'
-      const redirectUrl = '/redirectUrl'
-      let requestIsRedirected = false
-      server.on('request', (request, response) => {
-        switch (request.url) {
-          case redirectUrl:
-            requestIsRedirected = true
-            response.end()
-            break
-          default:
-            handleUnexpectedURL(request, response)
-        }
-      })
-
-      let requestIsIntercepted = false
-      session.defaultSession.webRequest.onBeforeRequest(
-        (details, callback) => {
-          if (details.url === `${server.url}${requestUrl}`) {
-            requestIsIntercepted = true
-            // Disabled due to false positive in StandardJS
-            // eslint-disable-next-line standard/no-callback-literal
-            callback({
-              redirectURL: `${server.url}${redirectUrl}`
-            })
-          } else {
-            callback({
-              cancel: false
-            })
-          }
-        })
-
-      const urlRequest = net.request(`${server.url}${requestUrl}`)
-
-      urlRequest.on('response', (response) => {
-        expect(response.statusCode).to.equal(200)
-        response.pause()
-        response.on('data', (chunk) => {
-        })
-        response.on('end', () => {
-          expect(requestIsRedirected).to.be.true('The server should receive a request to the forward URL')
-          expect(requestIsIntercepted).to.be.true('The request should be intercepted by the webRequest module')
-          done()
-        })
-        response.resume()
-      })
-      urlRequest.end()
-    })
-
-    it('should to able to create and intercept a request using a custom session object', (done) => {
-      const requestUrl = '/requestUrl'
-      const redirectUrl = '/redirectUrl'
-      const customPartitionName = 'custom-partition'
-      let requestIsRedirected = false
-      server.on('request', (request, response) => {
-        switch (request.url) {
-          case redirectUrl:
-            requestIsRedirected = true
-            response.end()
-            break
-          default:
-            handleUnexpectedURL(request, response)
-        }
-      })
-
-      session.defaultSession.webRequest.onBeforeRequest((details, callback) => {
-        expect.fail('Request should not be intercepted by the default session')
-      })
-
-      const customSession = session.fromPartition(customPartitionName, { cache: false })
-      let requestIsIntercepted = false
-      customSession.webRequest.onBeforeRequest((details, callback) => {
-        if (details.url === `${server.url}${requestUrl}`) {
-          requestIsIntercepted = true
-          // Disabled due to false positive in StandardJS
-          // eslint-disable-next-line standard/no-callback-literal
-          callback({
-            redirectURL: `${server.url}${redirectUrl}`
-          })
-        } else {
-          callback({
-            cancel: false
-          })
-        }
-      })
-
-      const urlRequest = net.request({
-        url: `${server.url}${requestUrl}`,
-        session: customSession
-      })
-      urlRequest.on('response', (response) => {
-        expect(response.statusCode).to.equal(200)
-        response.pause()
-        response.on('data', (chunk) => {
-        })
-        response.on('end', () => {
-          expect(requestIsRedirected).to.be.true('The server should receive a request to the forward URL')
-          expect(requestIsIntercepted).to.be.true('The request should be intercepted by the webRequest module')
-          done()
-        })
-        response.resume()
-      })
-      urlRequest.end()
-    })
-
-    it('should throw if given an invalid redirect mode', () => {
-      const requestUrl = '/requestUrl'
-      const options = {
-        url: `${server.url}${requestUrl}`,
-        redirect: 'custom'
-      }
-      expect(() => {
-        net.request(options)
-      }).to.throw('redirect mode should be one of follow, error or manual')
-    })
-
-    it('should throw when calling getHeader without a name', () => {
-      expect(() => {
-        net.request({ url: `${server.url}/requestUrl` }).getHeader()
-      }).to.throw(/`name` is required for getHeader\(name\)\./)
-
-      expect(() => {
-        net.request({ url: `${server.url}/requestUrl` }).getHeader(null)
-      }).to.throw(/`name` is required for getHeader\(name\)\./)
-    })
-
-    it('should throw when calling removeHeader without a name', () => {
-      expect(() => {
-        net.request({ url: `${server.url}/requestUrl` }).removeHeader()
-      }).to.throw(/`name` is required for removeHeader\(name\)\./)
-
-      expect(() => {
-        net.request({ url: `${server.url}/requestUrl` }).removeHeader(null)
-      }).to.throw(/`name` is required for removeHeader\(name\)\./)
-    })
-
-    it('should follow redirect when no redirect mode is provided', (done) => {
-      const requestUrl = '/301'
-      server.on('request', (request, response) => {
-        switch (request.url) {
-          case '/301':
-            response.statusCode = '301'
-            response.setHeader('Location', '/200')
-            response.end()
-            break
-          case '/200':
-            response.statusCode = '200'
-            response.end()
-            break
-          default:
-            handleUnexpectedURL(request, response)
-        }
-      })
-      const urlRequest = net.request({
-        url: `${server.url}${requestUrl}`
-      })
-      urlRequest.on('response', (response) => {
-        expect(response.statusCode).to.equal(200)
-        done()
-      })
-      urlRequest.end()
-    })
-
-    it('should follow redirect chain when no redirect mode is provided', (done) => {
-      const requestUrl = '/redirectChain'
-      server.on('request', (request, response) => {
-        switch (request.url) {
-          case '/redirectChain':
-            response.statusCode = '301'
-            response.setHeader('Location', '/301')
-            response.end()
-            break
-          case '/301':
-            response.statusCode = '301'
-            response.setHeader('Location', '/200')
-            response.end()
-            break
-          case '/200':
-            response.statusCode = '200'
-            response.end()
-            break
-          default:
-            handleUnexpectedURL(request, response)
-        }
-      })
-      const urlRequest = net.request({
-        url: `${server.url}${requestUrl}`
-      })
-      urlRequest.on('response', (response) => {
-        expect(response.statusCode).to.equal(200)
-        done()
-      })
-      urlRequest.end()
-    })
-
-    it('should not follow redirect when mode is error', (done) => {
-      const requestUrl = '/301'
-      server.on('request', (request, response) => {
-        switch (request.url) {
-          case '/301':
-            response.statusCode = '301'
-            response.setHeader('Location', '/200')
-            response.end()
-            break
-          case '/200':
-            response.statusCode = '200'
-            response.end()
-            break
-          default:
-            handleUnexpectedURL(request, response)
-        }
-      })
-      const urlRequest = net.request({
-        url: `${server.url}${requestUrl}`,
-        redirect: 'error'
-      })
-      urlRequest.on('error', (error) => {
-        expect(error.message).to.equal('Request cannot follow redirect with the current redirect mode')
-      })
-      urlRequest.on('close', () => {
-        done()
-      })
-      urlRequest.end()
-    })
-
-    it('should allow follow redirect when mode is manual', (done) => {
-      const requestUrl = '/redirectChain'
-      let redirectCount = 0
-      server.on('request', (request, response) => {
-        switch (request.url) {
-          case '/redirectChain':
-            response.statusCode = '301'
-            response.setHeader('Location', '/301')
-            response.end()
-            break
-          case '/301':
-            response.statusCode = '301'
-            response.setHeader('Location', '/200')
-            response.end()
-            break
-          case '/200':
-            response.statusCode = '200'
-            response.end()
-            break
-          default:
-            handleUnexpectedURL(request, response)
-        }
-      })
-      const urlRequest = net.request({
-        url: `${server.url}${requestUrl}`,
-        redirect: 'manual'
-      })
-      urlRequest.on('response', (response) => {
-        expect(response.statusCode).to.equal(200)
-        expect(redirectCount).to.equal(2)
-        done()
-      })
-      urlRequest.on('redirect', (status, method, url) => {
-        if (url === `${server.url}/301` || url === `${server.url}/200`) {
-          redirectCount += 1
-          urlRequest.followRedirect()
-        }
-      })
-      urlRequest.end()
-    })
-
-    it('should allow cancelling redirect when mode is manual', (done) => {
-      const requestUrl = '/redirectChain'
-      let redirectCount = 0
-      server.on('request', (request, response) => {
-        switch (request.url) {
-          case '/redirectChain':
-            response.statusCode = '301'
-            response.setHeader('Location', '/redirect/1')
-            response.end()
-            break
-          case '/redirect/1':
-            response.statusCode = '200'
-            response.setHeader('Location', '/redirect/2')
-            response.end()
-            break
-          case '/redirect/2':
-            response.statusCode = '200'
-            response.end()
-            break
-          default:
-            handleUnexpectedURL(request, response)
-        }
-      })
-      const urlRequest = net.request({
-        url: `${server.url}${requestUrl}`,
-        redirect: 'manual'
-      })
-      urlRequest.on('response', (response) => {
-        expect(response.statusCode).that.equal(200)
-        response.pause()
-        response.on('data', (chunk) => {
-        })
-        response.on('end', () => {
-          urlRequest.abort()
-        })
-        response.resume()
-      })
-      urlRequest.on('close', () => {
-        expect(redirectCount).to.equal(1)
-        done()
-      })
-      urlRequest.on('redirect', (status, method, url) => {
-        if (url === `${server.url}/redirect/1`) {
-          redirectCount += 1
-          urlRequest.followRedirect()
-        }
-      })
-      urlRequest.end()
-    })
-
-    it('should throw if given an invalid session option', (done) => {
-      const requestUrl = '/requestUrl'
-      try {
-        const urlRequest = net.request({
-          url: `${server.url}${requestUrl}`,
-          session: 1
-        })
-
-        // eslint-disable-next-line
-        urlRequest
-      } catch (exception) {
-        done()
-      }
-    })
-
-    it('should to able to create and intercept a request using a custom partition name', (done) => {
-      const requestUrl = '/requestUrl'
-      const redirectUrl = '/redirectUrl'
-      const customPartitionName = 'custom-partition'
-      let requestIsRedirected = false
-      server.on('request', (request, response) => {
-        switch (request.url) {
-          case redirectUrl:
-            requestIsRedirected = true
-            response.end()
-            break
-          default:
-            handleUnexpectedURL(request, response)
-        }
-      })
-
-      session.defaultSession.webRequest.onBeforeRequest((details, callback) => {
-        expect.fail('Request should not be intercepted by the default session')
-      })
-
-      const customSession = session.fromPartition(customPartitionName, {
-        cache: false
-      })
-      let requestIsIntercepted = false
-      customSession.webRequest.onBeforeRequest((details, callback) => {
-        if (details.url === `${server.url}${requestUrl}`) {
-          requestIsIntercepted = true
-          callback({
-            redirectURL: `${server.url}${redirectUrl}`
-          })
-        } else {
-          callback({
-            cancel: false
-          })
-        }
-      })
-
-      const urlRequest = net.request({
-        url: `${server.url}${requestUrl}`,
-        partition: customPartitionName
-      })
-      urlRequest.on('response', (response) => {
-        expect(response.statusCode).to.equal(200)
-        response.pause()
-        response.on('data', (chunk) => {
-        })
-        response.on('end', () => {
-          expect(requestIsRedirected).to.be.true('The server should receive a request to the forward URL')
-          expect(requestIsIntercepted).to.be.true('The request should be intercepted by the webRequest module')
-          done()
-        })
-        response.resume()
-      })
-      urlRequest.end()
-    })
-
-    it('should throw if given an invalid partition option', (done) => {
-      const requestUrl = '/requestUrl'
-      try {
-        const urlRequest = net.request({
-          url: `${server.url}${requestUrl}`,
-          partition: 1
-        })
-
-        // eslint-disable-next-line
-        urlRequest
-      } catch (exception) {
-        done()
-      }
-    })
-
-    it('should be able to create a request with options', (done) => {
-      const requestUrl = '/'
-      const customHeaderName = 'Some-Custom-Header-Name'
-      const customHeaderValue = 'Some-Customer-Header-Value'
-      server.on('request', (request, response) => {
-        switch (request.url) {
-          case requestUrl:
-            expect(request.method).to.equal('GET')
-            expect(request.headers[customHeaderName.toLowerCase()]).to.equal(customHeaderValue)
-            response.statusCode = 200
-            response.statusMessage = 'OK'
-            response.end()
-            break
-          default:
-            handleUnexpectedURL(request, response)
-        }
-      })
-
-      const serverUrl = url.parse(server.url)
-      const options = {
-        port: serverUrl.port,
-        hostname: '127.0.0.1',
-        headers: {}
-      }
-      options.headers[customHeaderName] = customHeaderValue
-      const urlRequest = net.request(options)
-      urlRequest.on('response', (response) => {
-        expect(response.statusCode).to.be.equal(200)
-        response.pause()
-        response.on('data', (chunk) => {
-        })
-        response.on('end', () => {
-          done()
-        })
-        response.resume()
-      })
-      urlRequest.end()
-    })
-
-    it('should be able to pipe a readable stream into a net request', (done) => {
-      const nodeRequestUrl = '/nodeRequestUrl'
-      const netRequestUrl = '/netRequestUrl'
-      const bodyData = randomString(kOneMegaByte)
-      let netRequestReceived = false
-      let netRequestEnded = false
-      server.on('request', (request, response) => {
-        switch (request.url) {
-          case nodeRequestUrl:
-            response.write(bodyData)
-            response.end()
-            break
-          case netRequestUrl:
-            netRequestReceived = true
-            let receivedBodyData = ''
-            request.on('data', (chunk) => {
-              receivedBodyData += chunk.toString()
-            })
-            request.on('end', (chunk) => {
-              netRequestEnded = true
-              if (chunk) {
-                receivedBodyData += chunk.toString()
-              }
-              expect(receivedBodyData).to.be.equal(bodyData)
-              response.end()
-            })
-            break
-          default:
-            handleUnexpectedURL(request, response)
-        }
-      })
-
-      const nodeRequest = http.request(`${server.url}${nodeRequestUrl}`)
-      nodeRequest.on('response', (nodeResponse) => {
-        const netRequest = net.request(`${server.url}${netRequestUrl}`)
-        netRequest.on('response', (netResponse) => {
-          expect(netResponse.statusCode).to.be.equal(200)
-          netResponse.pause()
-          netResponse.on('data', (chunk) => {})
-          netResponse.on('end', () => {
-            expect(netRequestReceived).to.be.true()
-            expect(netRequestEnded).to.be.true()
-            done()
-          })
-          netResponse.resume()
-        })
-        nodeResponse.pipe(netRequest)
-      })
-      nodeRequest.end()
-    })
-
-    it('should emit error event on server socket close', (done) => {
-      const requestUrl = '/requestUrl'
-      server.on('request', (request, response) => {
-        switch (request.url) {
-          case requestUrl:
-            request.socket.destroy()
-            break
-          default:
-            handleUnexpectedURL(request, response)
-        }
-      })
-      let requestErrorEventEmitted = false
-      const urlRequest = net.request(`${server.url}${requestUrl}`)
-      urlRequest.on('error', (error) => {
-        expect(error).to.be.an('Error')
-        requestErrorEventEmitted = true
-      })
-      urlRequest.on('close', () => {
-        expect(requestErrorEventEmitted).to.be.true()
-        done()
-      })
-      urlRequest.end()
-    })
-  })
-
   describe('IncomingMessage API', () => {
     it('response object should implement the IncomingMessage API', (done) => {
       const requestUrl = '/requestUrl'
